requirements.txt
# Dépendances Python globales pour le développement local
# Ces packages sont utiles pour développer et tester en dehors des conteneurs Docker

# Kafka
kafka-python==2.0.2

# Spark (pour développement local)
# pyspark==3.5.0  # Décommentez si vous voulez utiliser Spark localement

# MLflow
mlflow==2.8.1

# FastAPI (pour développement local)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Base de données
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Redis
redis==5.0.1

# Utilitaires
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0

# Tests (optionnel)
pytest==7.4.3
pytest-asyncio==0.21.1

